Upload TypeScript logo
🎨 Uploaded the TypeScript logo (typescript.png) to the "logos" folder in the repository. The logo will be used to represent TypeScript technology in the README.md file and any other relevant sections.

The TypeScript logo is an essential visual element to showcase the technology I use in my projects. It will be incorporated into the "Technologies I Use" section of the README.md file, visually representing my proficiency with TypeScript among other technologies.

This addition enhances the overall presentation of my GitHub profile and reinforces my expertise in TypeScript, contributing to a more engaging and informative experience for visitors.

The logo file (typescript.png) is now available in the "logos" folder for easy access and organization, ensuring it can be utilized in future updates to the README.md file and other relevant sections.

I'll proceed to add the TypeScript logo to the "Technologies I Use" section in the README.md file to reflect its addition to the list of technologies I work with. 🌟