Fix logo width and alt attribute for new Technologies and Tools
🚀 Made adjustments to the logos of the recently added Technologies and Tools in the "Technologies and Tools I Use" section of the README.md file for better display and accessibility.

In this update, I fixed the width of the Tailwind CSS logo to ensure uniformity with the other logos. Additionally, I enhanced the alt attribute for the following logos:

- Tailwind CSS
- Adobe Illustrator
- Adobe Photoshop
- Figma
- EmailJS

With these improvements, the logos now render consistently in the document, creating a visually appealing and well-structured presentation. Moreover, the revised alt attributes provide more descriptive and accessible information, making it easier for users with disabilities to understand the purpose of each logo.

I'm committed to maintaining an inclusive and user-friendly experience, and these adjustments align with that goal. By ensuring the proper display and accessibility of the logos, visitors can seamlessly explore my proficiency in various technologies and tools as a Front End Developer.

Let's continue to enhance our web development journey together! 🌐💻